Moved experimentHelpers to lib
